Add pseudo code in README
This is a fix, some words got jumbled with the text instruction

Created an outline for what the code will be nested.
This is to help in understanding how the code will work.
The code written is just an outline and not an official syntax of any language (That I know of.)